% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_reflectance}
\alias{add_reflectance}
\title{Add Reflectance column to LAS if it is missing for RIEGL vz400i}
\usage{
add_reflectance(las)
}
\arguments{
\item{las}{`LAS` object from `lidR` package representing an individually
segmented tree containing an `Intensity` column representing 16-bit range-
corrected amplitude from RIEGL vz400i Terrestrial Lidar Scanner}
}
\description{
Function to provide relative Reflectance for RIEGL vz400i. Lidar prediction
of crown scorch is based on relative range-corrected
reflectance relative to a white reference object orthonormal to
scanner. Raw range-corrected amplitudes from RIEGL vz400i are linearly
correlated to relative intensity which usually ranges from -20 dB to 0 dB
}
\examples{
library(lidR)
library(CrownScorchTLS)
las_file = system.file('extdata', 'tree_005.laz', package = 'CrownScorchTLS')
las = readLAS(las_file)
las = add_reflectance(las)
colnames(las@data)
}
