% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_histogram}
\alias{get_histogram}
\title{Generate histogram of Reflectance for prediction wiht Random Forests}
\usage{
get_histogram(las, breaks = seq(-20, 0, by = 0.2))
}
\arguments{
\item{las}{`LAS` object from `lidR` package representing an individually
segmented tree containing a `Reflectance` column representing relative
reflectance from from RIEGL vz400i Terrestrial Lidar Scanner. See `add_reflectance()`}

\item{breaks}{sequence of breaks for histograms, default from Cannon et al. 2025.}
}
\description{
Generates a histogram of Reflectance intensities for prediction with
Random Forests. Histogram breaks can be defined.
}
\examples{
library(lidR)
library(CrownScorchTLS)
las_file = system.file('extdata', 'tree_005.laz', package = 'CrownScorchTLS')
las = readLAS(las_file)
las = add_reflectance(las)
histogram = get_histogram(las)
plot(density ~ intensity, data=histogram, xlab='Reflectance (dB)', type='l')
}
