% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_stem}
\alias{remove_stem}
\title{Remove tree bole from `LAS`}
\usage{
remove_stem(las)
}
\arguments{
\item{las}{`LAS` object from `lidR` package representing an individually
segmented tree}
}
\value{
LAS object with stem removed
}
\description{
This function identifies and removes tree boles using the
`TreeLS` package available at https://github.com/tiagodc/treels
}
\examples{
\dontrun{
library(lidR)
library(CrownScorchTLS)
las_file = system.file('extdata', 'tree_005.laz', package = 'CrownScorchTLS')
las = readLAS(las_file)
plot(las)
crown_only = remove_stem(las)
plot(crown_only)
}
}
