% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeLS_stemPoints_methods.R
\name{stemPoints}
\alias{stemPoints}
\title{Stem points classification}
\usage{
stemPoints(las, method = stm.hough())
}
\arguments{
\item{las}{\code{\link[lidR:LAS]{LAS}} object.}

\item{method}{Function to classify stems. Default: \code{\link{stm.hough}}.}
}
\value{
\code{\link[lidR:LAS]{LAS}} object.
}
\description{
Classify stem points of all trees in a \strong{normalized}
point cloud. Stem denoising methods are prefixed by \code{stm}.
This file includes code derived from the TreeLS package by Tiago de Conto
Original source: https://github.com/tiagodc/TreeLS
License: GPL-3
The code below is copied and adapted from TreeLS::stemPoints for the purpose
of maintaining CRAN compatibility. All modifications are clearly documented.
}
\note{
This function includes code derived from TreeLS::stemPoints
  (GPL-3 license). See source for details.
}
\references{
Carvalho, T. (2017). TreeLS: Tools for Terrestrial LiDAR in R.
  GitHub: https://github.com/tiagodc/TreeLS
}
